Separate circularDialMenu into own file
Update dependencies
