+++ lib/Musical-Terms/Definition-Page.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:circular_menu/circular_menu.dart';
-import 'dart:math' as math;
+import 'package:artifact/circular_dial_menu.dart';
 import 'package:artifact/Musical-Terms/Term.dart';
 import '../Composers/Composers.dart';
 import '../Musical-Works/Musical-Works.dart';
@@ -14,56 +13,8 @@ class DefScreen extends StatelessWidget {
 
   final Term term;
 
-  // Globally changes the strength/contrast of a color for an icon
-  final int iconColor = 300;
-
   @override
   Widget build(BuildContext context) {
-    // returns the circular spanning button for module selection
-    CircularMenu buildCircularDialMainMenu() {
-      return CircularMenu(
-        toggleButtonBoxShadow: const [],
-        startingAngleInRadian: 7 * math.pi / 6,
-        endingAngleInRadian: 11 * math.pi / 6,
-        items: [
-          CircularMenuItem(
-            boxShadow: const [],
-            color: Colors.yellow[iconColor],
-            icon: Icons.person,
-            iconColor: Colors.black,
-            onTap: () => Navigator.push(
-                context,
-                MaterialPageRoute(
-                    builder: (context) => const ComposersScreen())),
-          ),
-          CircularMenuItem(
-            boxShadow: const [],
-            color: Colors.blue[iconColor],
-            icon: Icons.music_note,
-            iconColor: Colors.black,
-            onTap: () => Navigator.push(context,
-                MaterialPageRoute(builder: (context) => const TermsScreen())),
-          ),
-          CircularMenuItem(
-            boxShadow: const [],
-            color: Colors.green[iconColor],
-            icon: Icons.quiz,
-            iconColor: Colors.black,
-            onTap: () => Navigator.push(context,
-                MaterialPageRoute(builder: (context) => const QuizzesScreen())),
-          ),
-          CircularMenuItem(
-            boxShadow: const [],
-            color: Colors.red[iconColor],
-            icon: Icons.book,
-            iconColor: Colors.black,
-            onTap: () => Navigator.push(context,
-                MaterialPageRoute(builder: (context) => const WorksScreen())),
-          ),
-        ],
-      );
-    }
-
     return Scaffold(
       body: SafeArea(
         child: Column(children: <Widget>[
@@ -98,7 +49,7 @@ class DefScreen extends StatelessWidget {
         ]),
       ),
       floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      floatingActionButton: buildCircularDialMainMenu(),
+      floatingActionButton: CircularDialMenu.build(context),
       bottomNavigationBar: BottomAppBar(
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
